separated buyer views
