+++ ecommerce_project/myapp/views/__init__.py
@@ -4,3 +4,4 @@ from .cart_views import *
 from .product_views import *
 from .review_views import *
 from .order_views import *
+from .buyer_views import *
